requirements.txt
numpy
pandas
matplotlib
seaborn
networkx
faiss-cpu
sentence-transformers
scikit-learn